refactor: split app persistence and recovery
